add content not found and move logic to ErrroTab
